# docker/backend/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: ThaliumX Exchange API
  description: |
    # ThaliumX Exchange Platform API
    
    A comprehensive cryptocurrency exchange and financial services platform API.
    
    ## Authentication
    
    Most endpoints require authentication using JWT tokens. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limiting
    
    - Standard endpoints: 100 requests per 15 minutes
    - Financial endpoints: 50 requests per 15 minutes
    - Authentication endpoints: 10 requests per 15 minutes
    
    ## Error Handling
    
    All errors follow a consistent format:
    ```json
    {
      "success": false,
      "error": {
        "code": "ERROR_CODE",
        "message": "Human readable message"
      }
    }
    ```
    
  version: 1.0.0
  contact:
    name: ThaliumX Support
    email: support@thaliumx.com
    url: https://thaliumx.com
  license:
    name: Proprietary
    url: https://thaliumx.com/terms
servers:
  - url: https://api.thaliumx.com
    description: Production server
  - url: https://staging-api.thaliumx.com
    description: Staging server
  - url: http://localhost:3002
    description: Development server
...
paths:
...
  /api/exchange/orderbook/{symbol}:
    get:
      tags:
        - Exchange
      summary: Get order book
      description: Returns the order book for a trading pair
      operationId: getOrderBook
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
          example: BTC-USDT
        - name: depth
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Order book data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderBook'
  /api/exchange/orders:
    post:
      tags:
        - Exchange
      summary: Place an order
      description: Places a new trading order
      operationId: placeOrder
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderRequest'
      responses:
        '201':
          description: Order placed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Invalid order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Insufficient balance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/cex/market-data/{symbol}:
    get:
      tags:
        - Exchange
      summary: Get market data
      description: Returns detailed market data for a trading pair
      operationId: getMarketData
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Market data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarketData'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
...
    OrderBook:
      type: object
      properties:
        symbol:
          type: string
        bids:
          type: array
          items:
            type: array
            items:
              type: string
        asks:
          type: array
          items:
            type: array
            items:
              type: string
        timestamp:
          type: string
          format: date-time
    OrderRequest:
      type: object
      required:
        - symbol
        - side
        - type
        - quantity
      properties:
        symbol:
          type: string
        side:
          type: string
          enum: [buy, sell]
        type:
          type: string
          enum: [market, limit, stop-loss, stop-limit]
        quantity:
          type: string
        price:
          type: string
        stopPrice:
          type: string
        timeInForce:
          type: string
          enum: [GTC, IOC, FOK]
          default: GTC
    OrderResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            orderId:
              type: string
            symbol:
              type: string
            side:
              type: string
            type:
              type: string
            quantity:
              type: string
            price:
              type: string
            status:
              type: string
              enum: [pending, open, filled, partially_filled, cancelled]
            createdAt:
              type: string
              format: date-time
...
    MarketData:
      type: object
      properties:
        symbol:
          type: string
        lastPrice:
          type: string
        volume24h:
          type: string
        change24h:
          type: string
        high24h:
          type: string
        low24h:
          type: string
        openInterest:
          type: string
        fundingRate:
          type: string